Bisected and Fixed.
Detail. If EEPROM is witten to continuously it will mess with ISRs
on a attiny and cause millis and micros to not be stable.
But was introduced by refactoring the storage of engine hours.
Now need to check that the 328p code is not in the same state on
the device where it was flashed as that could break the EEPROM
on the chip.